{{cookiecutter.repo_name}}/requirements.txt
bottle
{% if cookiecutter.use_db == "y" -%}
bottle-sqlalchemy==0.4.3
psycopg2
{%- endif -%}
python-decouple
